fix crash on server mayhaps
